Rename getItemsHelper.ts to getItems.ts
