+++ src/plugins/chromecastPlayer/plugin.js
@@ -6,7 +6,7 @@ import castSenderApiLoader from './castSenderApi';
 import ServerConnections from '../../components/ServerConnections';
 import alert from '../../components/alert';
 import Events from '../../utils/events.ts';
-import { getItems } from '../../utils/jellyfin-apiclient/getItemsHelper.ts';
+import { getItems } from '../../utils/jellyfin-apiclient/getItems.ts';
 
 // Based on https://github.com/googlecast/CastVideos-chrome/blob/master/CastVideos.js
 
@@ -482,7 +482,7 @@ function getItemsForPlayback(apiClient, query) {
         query.ExcludeLocationTypes = 'Virtual';
         query.EnableTotalRecordCount = false;
 
-        // call getItems from getItemsHelper instead of apiClient.getItems()
+        // call getItems from jellyfin-apiclient/getItems.ts instead of apiClient.getItems()
         // to split up into multiple requests if necessary (URL might get too long)
         return getItems(apiClient, userId, query);
     }
